Move tests to a different namespace
