+++ .travis.yml
@@ -28,8 +28,8 @@ before_install:
 
 install:
     - composer update --prefer-dist
-    - src/Tests/Application/bin/console doctrine:schema:create --env=test -vvv --no-interaction
-    - src/Tests/Application/bin/console fos:elastica:reset --env=test -vvv --no-interaction
+    - tests/Application/bin/console doctrine:schema:create --env=test -vvv --no-interaction
+    - tests/Application/bin/console fos:elastica:reset --env=test -vvv --no-interaction
 
 script:
     - bin/behat --strict -vvv --no-interaction
