+++ composer.json
@@ -43,6 +43,9 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "autoload": {
-        "psr-4": { "Lakion\\SyliusElasticSearchBundle\\": "src/" }
+        "psr-4": {
+            "Lakion\\SyliusElasticSearchBundle\\": "src/",
+            "Tests\\Lakion\\SyliusElasticSearchBundle\\": "tests/"
+        }
     }
 }
