+++ phpunit.xml.dist
@@ -7,12 +7,12 @@
          bootstrap="vendor/autoload.php">
     <testsuites>
         <testsuite name="Sylius Test Suite">
-            <directory>src/Tests</directory>
+            <directory>tests/</directory>
         </testsuite>
     </testsuites>
 
     <php>
-        <server name="KERNEL_CLASS_PATH" value="src/Tests/Application/app/AppKernel.php" />
+        <server name="KERNEL_CLASS_PATH" value="tests/Application/app/AppKernel.php" />
         <server name="IS_DOCTRINE_ORM_SUPPORTED" value="true" />
     </php>
 </phpunit>
