+++ tests/Behat/Resources/suites/shop/domain/product/searching_products.yml
