+++ tests/Behat/Resources/suites/shop/ui/product/searching_products.yml
